bigcontest2025/pyproject.toml
[project]
name = "shcard_2025_bigcontest_chatbot"
version = "0.1.0"
description = "신한카드 가맹점 추천 챗봇"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.38.0",
    "google-generativeai>=0.8.0",
    "pandas>=2.2.0",
    "mcp>=1.13.1",
    "fastmcp>=2.11.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.1.0",
    "pillow>=10.0.0",
    "asyncio>=4.0.0",
]
